+++ ext/land_sim_vis/leaderboard/leaderboard.jl
@@ -205,7 +205,7 @@ function compute_monthly_leaderboard(
         months =
             Dates.month.(
                 Dates.DateTime(sim_var.attributes["start_date"]) .+
-                Dates.Second.(times)
+                Dates.Second.(times),
             )
         months_split, sim_vec_split, rmse_vec_split, bias_vec_split =
             partition_by_val(12, months, sim_vec, rmse_vec, bias_vec)